+++ RELEASE_NOTES.md
@@ -8,6 +8,12 @@
   This initial seed only supports `${xmlunit.ignore}` which can be
   used to make XMLUnit ignore the element containing this text.
 
+* fixed the message when `CompareConstraint` or
+  `ValidationConstraints` (both NUnit 2.x and 3.x) as well as
+  `EvaluateXPathConstraint` or `HasXPathConstraint` (only the NUnit
+  3.x versions) pass but the assertion failes because the constraint
+  itself was wrapped in a `Not` constraint.
+
 ## XMLUnit.NET 2.5.1 - /Released 2017-10-20/
 
 * elements that only differed in namespace prefix resulted in a false